requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coverage==4.5.4
coveralls==1.11.1
cryptography==2.9.2
Django==1.8.19
django-autofixture==0.12.1
django-mysql==1.0.9
docopt==0.6.2
enum34==1.1.10
idna==2.9
ipaddress==1.0.23
MySQL-python==1.2.5
pycparser==2.20
pyOpenSSL==19.1.0
python-dateutil==2.8.1
requests==2.23.0
six==1.14.0
travis==0.0.3
urllib3==1.25.9
mock==3.0.5
pytz==2019.3
ipdb==0.13.2
memory-profiler==0.57.0
factory-boy==2.12.0
